% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_gaus}
\alias{C_gaus}
\title{A random variable sampler using Rcpp}
\usage{
C_gaus(n)
}
\arguments{
\item{N}{the number of samples}
}
\value{
a random sample of size \code{n}
}
\description{
A random variable sampler using Rcpp
}
\examples{
\dontrun{
rnC = C_gaus(100)
}
}
