% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_mean}
\alias{C_mean}
\title{compute the mean of a vector}
\usage{
C_mean(X)
}
\arguments{
\item{N}{the number of samples}
}
\value{
a random sample of size \code{n}
\dontrun{
rnC = C_mean(c(1,2,3,4,5))
}
}
\description{
A random variable sampler using Rcpp
}
