% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMonte_antithetic.R
\name{RMonte_antithetic}
\alias{RMonte_antithetic}
\title{mediation analysis using R}
\usage{
RMonte_antithetic(FUN, low, up, N)
}
\arguments{
\item{N}{the number of samples}

\item{thin}{the number of between-sample random numbers}
}
\value{
a random sample of size \code{n}
}
\description{
mediation analysis using R
}
\examples{
\dontrun{
RMonte_antithetic(exp,0,1,10)
}
}
