% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMonte_stratify.R
\name{RMonte_stratify}
\alias{RMonte_stratify}
\title{mediation analysis using R}
\usage{
RMonte_stratify(FUN, low, up, N, layer)
}
\arguments{
\item{N}{the number of samples}

\item{thin}{the number of between-sample random numbers}
}
\value{
a random sample of size \code{n}
}
\description{
mediation analysis using R
}
\examples{
\dontrun{
RMonte_stratify(exp,0,1,10,1)
}
}
