% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestVariance.R
\name{TestVariance}
\alias{TestVariance}
\title{mediation analysis using R}
\usage{
TestVariance(TESTFUN, M, ...)
}
\arguments{
\item{TESTFUN}{the number of samples}

\item{M}{the number of between-sample random numbers}

\item{...}{hhh}
}
\value{
a random sample of size \code{n}
}
\description{
mediation analysis using R
}
\examples{
\dontrun{
TestVariance(RMonte,10,exp,1,5,100)
}
}
