% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{C_poly}
\alias{C_poly}
\title{generate a random variable from an exp-family distribution}
\usage{
C_poly(COEF_h, COEF, sigma, L, low, up)
}
\arguments{
\item{COEF_h}{polynomial of hx}

\item{COEF}{polynomial of the fucntion in exp item}

\item{sigma}{variance of the random walk}

\item{L}{length of the markov chain}

\item{low}{lower bound of the support set}

\item{up}{up bound of the support set}
}
\value{
a random sample of size \code{n}
}
\description{
generate a random variable from an exp-family distribution by MCMC
}
