% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cboot.R
\name{Cboot}
\alias{Cboot}
\title{Cboot}
\usage{
Cboot(DATA, sims = 10, Method = "Norm", alpha = 0.05)
}
\arguments{
\item{DATA}{a vector which contain the data}

\item{sims}{the number of simulations}

\item{Method}{algorithm for computing confidence interval. sohuld be Norm, perc, basic, stud}

\item{alpha}{1 - confidence level}
}
\value{
a random sample of size \code{n}
}
\description{
bootstrap estimation of mean of a vector
}
\examples{
\dontrun{
Cboot(1:100,10,'Norm',0.05)
}
}
