% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMonte.R
\name{RMonte}
\alias{RMonte}
\title{RMonte}
\usage{
RMonte(FUN, low, up, N, layer = 1, Method = "None")
}
\arguments{
\item{FUN}{the function want to integrate}

\item{low}{the lower bound of the integration interval}

\item{up}{the up bound of the integration interval}

\item{N}{time of simulations}

\item{layer}{the layer of stratification}

\item{Method}{the variance-reduction method,None,control,Antithetic}
}
\value{
the value of integration
}
\description{
Monte Carlo integration
}
\examples{
\dontrun{
RMonte(exp,0,1,10)
}
}
