% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestVariance.R
\name{TestVariance}
\alias{TestVariance}
\title{TestVariance}
\usage{
TestVariance(TESTFUN, M, ...)
}
\arguments{
\item{TESTFUN}{the number of samples}

\item{M}{the number of simulations}

\item{...}{other parameters}
}
\value{
real value number.
}
\description{
compute the variance of an estimation
}
\examples{
\dontrun{
TestVariance(RMonte,10,exp,1,5,100)
}
}
